Major ContentServiceImpl tests changes

+++ nmaas-platform/src/main/java/net/geant/nmaas/portal/service/impl/ContentServiceImpl.java
@@ -11,8 +11,11 @@ import java.util.Optional;
 
 public class ContentServiceImpl implements net.geant.nmaas.portal.service.ContentService {
 
-    @Autowired
     ContentRepository contentRepo;
+
+    public ContentServiceImpl(ContentRepository repository){
+        this.contentRepo = repository;
+    }
 
     @Override
     public Optional<Content> findByName(String name){
@@ -45,6 +48,7 @@ public class ContentServiceImpl implements net.geant.nmaas.portal.service.Conten
         }
 
         contentRepo.saveAndFlush(content);
+
     }
 
     @Override
@@ -52,20 +56,23 @@ public class ContentServiceImpl implements net.geant.nmaas.portal.service.Conten
         checkParam(content);
         checkParam(content.getId());
 
+        if(!contentRepo.existsById(content.getId())){
+            throw new ProcessingException("Content (id=" + content.getId() + ") does not exists.");
+        }
         contentRepo.delete(content);
     }
 
-    public void checkParam(Long id) {
+    private void checkParam(Long id) {
         if(id == null)
             throw new IllegalArgumentException("id is null");
     }
 
-    public void checkParam(String name) {
+    private void checkParam(String name) {
         if(name == null)
             throw new IllegalArgumentException("name is null");
     }
 
-    public void checkParam(Content content) {
+    private void checkParam(Content content) {
         if(content == null)
             throw new IllegalArgumentException("content is null");
     }
